Struct linked_tail_list::TailList

pub struct TailList<T> {
    // some fields omitted
}

A specialized linked list (see the module documentation).

Methods

impl<T> TailList<T>

fn new() -> TailList<T>

Creates a new empty list.

fn push(&mut self, val: T)

Pushed a new element to the front of the list.

fn cursor<'node>(&'node mut self) -> Cursor<'node, T>

Returns a cursor over all elements in this list.