Struct linked_tail_list::ValRef

pub struct ValRef<'node, T: 'node> {
    // some fields omitted
}

A ValRef is a passive item, which provides mutable access to a single node.

Methods

impl<'node, T: 'node> ValRef<'node, T>

fn insert_before(&mut self, val: T) -> ValRef<'node, T>

Inserts a new element before this element and returns a ValRef to the newly inserted element.

fn insert_after(&mut self, val: T) -> ValRef<'node, T>

Inserts a new element after this element and returns a ValRef to the newly inserted element.

fn remove(self) -> T

Removes this element from the list and returns it's value.