Struct linked_tail_list::Cursor

pub struct Cursor<'node, T: 'node> {
    // some fields omitted
}

A Cursor is an iterator over a node and it's tail. It is an active item, which owns the next node it would return.

Due to the design of rust's Iterator trait, Cursor cannot implement Iterator.

Methods

impl<'node, T: 'node> Cursor<'node, T>

fn next<'tail>(&'tail mut self) -> Option<TailValRef<'node, 'tail, T>>

(Optionally) returns the next element of this cursor.

This cursor is unusable as long as the 'tail lifetime is still referenced.